% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Data Cleaning for everything else}
\usage{
eq_clean_data(input_data)
}
\arguments{
\item{input_data}{raw earthquake data}
}
\value{
cleaned data, new column for DATE, numeric for DEATHS, LATITUDE, LONGITUDE
}
\description{
Data Cleaning for everything else
}
\examples{
\dontrun{
eq_clean_data(eq_data)
}
}
