% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Remove country name from the LOCATION_NAME}
\usage{
eq_location_clean(LOCATION_NAME)
}
\arguments{
\item{LOCATION_NAME}{single string that contains country and location}
}
\value{
string with location only, country is removed
}
\description{
Remove country name from the LOCATION_NAME
}
\examples{
\dontrun{
location_list <- c("MEXICO: SAN ANDRES TUXTLA, TUXTEPEC", "MEXICO: MEXICALI, BAJA CALIFORNIA", "MEXICO: ACAPULCO")
eq_location_clean(location_list)
}

}
