% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{eq_map}
\alias{eq_map}
\title{Maps the epicenters}
\usage{
eq_map(input_data, annot_col = "DATE")
}
\arguments{
\item{input_data}{Earthquake data}

\item{annot_col}{column to show (in this case, DATE)}
}
\value{
Earthquake data Clean
}
\description{
Maps the epicenters
}
\examples{
\dontrun{
file_name <- system.file("eq_data", "extdata/earthquakes.txt", package = "EQ")
eq_data   <- file_name                                                          \%>\%
             readr::read_delim(txt, delim = "\t")                               \%>\%
             eq_clean_data()                                                    \%>\%
             dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
             dplyr::mutate(popup_text = eq_create_label(.))                     \%>\%
             eq_map(annot_col = "popup_text")
}
}
