% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\docType{data}
\name{geom_point}
\alias{geom_point}
\title{Plot timeline earthquakes}
\format{
An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
geom_point
}
\value{
mapped aesthetics for the Geom layer
}
\description{
Plot timeline earthquakes
}
\keyword{datasets}
