% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{geom_timelinelabel}
\alias{geom_timelinelabel}
\title{Plot timeline for earthquakes}
\usage{
geom_timelinelabel(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  show.legend = NA,
  na.rm = FALSE,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{mapping parameter, leave as NULL}

\item{data}{data parameter, leave as NULL}

\item{stat}{stat parameter}

\item{position}{position parameter}

\item{show.legend}{show.legend parameter}

\item{na.rm}{na.rm parameter}

\item{inherit.aes}{inherit.aes parameter}

\item{...}{other parameters}
}
\description{
Plot timeline for earthquakes
}
\examples{
\dontrun{
geom_timelinelabel(eq_data)
}
}
