% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\docType{data}
\name{geo_timeline_point}
\alias{geo_timeline_point}
\title{text annotation}
\format{
An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
geo_timeline_point
}
\description{
text annotation
}
\keyword{datasets}
